remove httpClientConfiguration
